+++ model_navigator/framework_api/package_descriptor.py
@@ -51,7 +51,7 @@ from model_navigator.framework_api.utils import (
 from model_navigator.model import Format
 from model_navigator.utils.environment import get_env, get_git_info
 
-NAV_PACKAGE_FORMAT_VERSION = "0.1.1"
+NAV_PACKAGE_FORMAT_VERSION = "0.1.2"
 
 
 class PackageDescriptor: